nihals version3

+++ routes/api/projects.js
@@ -13,7 +13,7 @@ router.get('/', async (req,res) => {
 })
 
 
-// Create a project
+// Create a projectt
 router.post('/', async (req,res) => {
    try {
     const isValidated = validator.createValidation(req.body)
@@ -48,11 +48,15 @@ router.put('/:id', async (req,res) => {
  // Delete a project
  router.delete('/:id',async(req,res)=>{
     try {
-        const id = req.params.id
-        const deletedProject = await Project.findByIdAndRemove(id)
-        if(!deletedProject) return res.status(400).send({ error: 'Project does not exist'})
-        res.json({msg:'Project was deleted successfully', data: deletedProject})
+        if(ObjectId.isValid(req.params.id)){
+            const id = req.params.id
+            const deletedProject = await Project.findByIdAndRemove(id)
+            if(!deletedProject) return res.status(400).send({ error: 'Project does not exist'})
+            res.json({msg:'Project was deleted successfully', data: deletedProject})
+       }else{
+            return res.status(404).send({error: 'Project does not exist'})
        }
+    }
        catch(error) {
            console.log(error)
            return res.status(400).send('not a project id')
